+++ positron/Drawables/IUpdateSync.cs
@@ -5,10 +5,10 @@ namespace positron
 {
     public class UpdateEventArgs : EventArgs
     {
-        protected double _Time;
-        public double Time { get { return _Time; } set { _Time = value; } }
+        protected float _Time;
+        public float Time { get { return _Time; } set { _Time = value; } }
 		public int _dbg_list_idx;
-        public UpdateEventArgs(double time, int dbg_list_idx = 0)
+        public UpdateEventArgs(float time, int dbg_list_idx = 0)
         {
             _Time = time;
 			_dbg_list_idx = dbg_list_idx;
@@ -19,6 +19,6 @@ namespace positron
     {
         //event UpdateEventHandler UpdateEvent;
 		List<KeyValuePair<object, UpdateEventHandler>> UpdateEventList { get; }
-        void Update(double time);
+        void Update(float time);
     }
 }
