+++ positron/IRenderable.cs
@@ -5,7 +5,7 @@ namespace positron
 	public interface IRenderable : IDisposable
 	{
         RenderSet Set { get; }
-        void Render(double time);
+        void Render(float time);
 		void Dispose();
 	}
 }
