+++ positron/MathUtil.cs
@@ -7,13 +7,13 @@ namespace positron
 		public static T Clamp<T>(T x, T hi, T lo) where T : System.IComparable<T> {
 			return x.CompareTo(hi) > 0 ? hi : x.CompareTo(lo) < 0 ? lo: x;
 		}
-		public static double Trapz(double x, double width, double offset)
+		public static float Trapz(float x, float width, float offset)
 		{
-			return MathUtil.Clamp(3.0 * width - Math.Abs(2.0 * x - offset), 2.0 * width, 0.0) * 0.5;
+			return MathUtil.Clamp(3.0f * width - Math.Abs(2.0f * x - offset), 2.0f * width, 0.0f) * 0.5f;
 		}
-		public static double Trapz(double x) // width: 1, offset: 3, period == 4
+		public static float Trapz(float x) // width: 1, offset: 3, period == 4
 		{
-			return MathUtil.Clamp(3.0 - Math.Abs(2.0 * x - 3.0), 2.0, 0.0) * 0.5;
+			return MathUtil.Clamp(3.0f - Math.Abs(2.0f * x - 3.0f), 2.0f, 0.0f) * 0.5f;
 		}
 		public static int Trapz(int x, int width, int offset)
 		{
